Add cast in debug message to avoid warnings about %d format specifier not matching a 64-bit size_t.
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1450 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/crypt3/cryptkeyrsa.c
@@ -395,7 +395,7 @@ int GWEN_Crypt_KeyRsa__ReadMpi(GWEN_DB_NODE *db,
     return GWEN_ERROR_GENERIC;
   }
   if (nscanned<1) {
-    DBG_INFO(GWEN_LOGDOMAIN, "Empty %s (%d)", dbName, nscanned);
+    DBG_INFO(GWEN_LOGDOMAIN, "Empty %s (%d)", dbName, (int)nscanned);
 #if 0
     if (mpi)
       gcry_mpi_release(mpi);
